// src/auth/interfaces/user.ts
export interface IUserLogin {
    username: string;
    name: string;
    lastname: string;
    token:    string;
    role:     string;
    _id:      string;
}



export interface IRegisterUser {
    name: string;
    lastname: string;
    username: string;
    password?: string;
    mail: string;
    role: string;
    devTitle: string;
    stack?: string | string[];
    image?: string;
    id?: string;

}

// Generated by https://quicktype.io

export interface IUsers {
    _id:      string;
    name:     string;
    lastname: string;
    stack: string[];
    devTitle: string;
}

export interface IUser {
    name: string;
    lastname: string;
    username: string;
    password?: string;
    mail: string;
    role: string;
    devTitle: string;
    stack?: string | string[];
    image?: string;
    id?: string;
}